[form] WIP: Streamline field exports and enhance translation keys
