--- packages/form/src/fields/array/context.ts
@@ -1,30 +1,0 @@
-import { getContext, setContext } from "svelte";
-
-import type { Schema, Validator } from "@/core/index.js";
-
-import type { FieldError, PossibleError } from "@/form/index.js";
-
-export interface ArrayContext<V extends Validator> {
-  canAdd: boolean;
-  addable: boolean;
-  orderable: boolean;
-  removable: boolean;
-  copyable: boolean;
-  errors: FieldError<PossibleError<V>>[];
-  key(index: number): number;
-  pushItem(itemSchema: Schema): void;
-  moveItemUp(index: number): void;
-  moveItemDown(index: number): void;
-  copyItem(index: number): void;
-  removeItem(index: number): void;
-}
-
-const ARRAY_CONTEXT = Symbol("array-context");
-
-export function getArrayContext<V extends Validator>(): ArrayContext<V> {
-  return getContext(ARRAY_CONTEXT);
-}
-
-export function setArrayContext<V extends Validator>(ctx: ArrayContext<V>) {
-  setContext(ARRAY_CONTEXT, ctx);
-}
