+++ packages/form/src/fields/object/exports.ts
@@ -1,3 +1,2 @@
-export * from './context.js'
 export { default as objectField } from "./object.svelte";
 export { default as objectPropertyField } from './object-property.svelte'
